[m88k] Fix problem with load
In some cases it happened that a physical register was used as the base
register for a load after instruction selection. However, this is not
valid MIR. The problem was caused returning the operand 1 of 'COPY $r2'
instead of using register of operand 0.
With the fix, the code is also slightly restructured.