[m88k] Legalize part of G_UITOFP
GUITOFP can be lowered to G_SITOFP. However, this requires
creating a new basic block, which the legalizer does not
support. The solution is to declare G_UITOFP as legal, and
lower the instruction before the instruction select.

Only i32 -> float/double is implemented. A i64 input requires a
different lowering.