[m88k] Re-implement overflow handling
Using a bit to represent the carry turned out to have some problems.
With this commit, the carry is represented by a virtual 32 bit register.
The code generation is changed to check that a carry-in is not destroyed
by instructions between the carry produce and the carry consumer.
Code to move the carry to a general purpose register and back is inserted
when the carry is destroyed.
As a side-effect, some patterns got easier.