[m88k] Generalize mapping of intrinsics
Assumes that all register arguments are 32 bit GPRs. If there are
exceptions to this, then a special check needs to be inserted. This
approach could be used for the other instructions, too, and could
reduce the code size of the register bank information a lot.

Also adds test cases and corrections for the llvm.returnaddress and
llvm.frameaddressintrinsics, which cannot be used without this
change.