[m88k] Rework RegisterBankInfo
For the m88k architecture, the register bank info is somewhat simple.
This change puts the instructions in different groups:

1. FP instructions with all operands having the same type
   Example: G_FADD
2. FP instructions with operands of different type
   Example: G_FPEXTEND
3. Instructions which may use different register banks
   Example: G_LOAD, G_INTRINSIC_TRUNC, COPY
4. Instructions which only use the common register bank
   Example: G_OR

The 1st and 4th case are handled as the default case. The instructions for
the 2nd case are grouped together. Only the instructions of the 3rd case
require case-by-case handling.

Most instructions fall into the default case, which simplifies coding a lot.