[PR99896] Generate getRegBankFromRegClass
Generating the mapping from a register class to a register bank
is complex:
 - there can be lots of register classes
 - the mapping may be ambiguos
    - a register class can span several register banks
      (e.g. a register class containing all registers)
    - the type information is not enough to decide which register
      bank to map to
      (e.g. a register class containing floating point and vector
      registers, and all register can represent a f64 value)

The approach taken here is to encode the register banks in an
array indexed by the ID of the register class. To save space, the
entries are packed into chunks of size 2^n.

See https://github.com/llvm/llvm-project/pull/99896